Rename folder diagram-types -> fhi-diagram
