+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram/diagram-type.model.ts
