+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram/diagram-types.ts
